bpf: Migrate cilium-migrate-map From C to Go
Currently initializing bpf programs and migrating maps
is done with a bash script. This creates a Go version of
the "cilium-migrate-map" program and refactors it to be
useful by a, similarly, migrated init program. Specifically,
giving feedback as to whether any maps even need migrating
(often they don't) thus allowing the init program to
skip the expensive loading and processing of ELF files.

Additionally, a Go version of the C command line program
is also added to the cilium container cli program to be used,
temporarily, by the init.sh program to ensure that the logic
of the new Go migrate program is sound.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>